requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
flexpolyline==0.1.0
geographiclib==2.0
geojson==3.1.0
geopy==2.4.1
googlemaps==4.10.0
idna==3.6
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.1
pyhocon==0.3.60
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
